Fixed crash if more than one kart is connected to the server (since the network kart screen will only show local karts now).

+++ src/states_screens/kart_selection.cpp
@@ -817,7 +817,8 @@ void KartSelectionScreen::updateKartStats(uint8_t widget_id,
     }
 }
 
-void KartSelectionScreen::updateKartWidgetModel(uint8_t widget_id,
+// ----------------------------------------------------------------------------
+void KartSelectionScreen::updateKartWidgetModel(int widget_id,
                 const std::string& selection,
                 const irr::core::stringw& selectionText)
 {
